DEMRUM-3394: Update per swiftformat and lint
